Feat : 인기 장소 내보내기, star 값 업데이트 Chore : inputZonecode 추가

+++ src/main/java/com/myspring/daengnyang/animal/vo/AnimalDetailVO.java
@@ -26,7 +26,7 @@ public class AnimalDetailVO {
     private String infield;
     private String outfield;
     private String addPay;
-    private Integer star;
+    private double star;
     private String tel;
     private String imgPath;
 }
