+++ src/main/java/com/myspring/daengnyang/animal/vo/AnimalReviewVO.java
@@ -15,6 +15,6 @@ public class AnimalReviewVO {
     private Integer memberNo;
     private String profileImg;
     private String review;
-    private Integer star;
+    private double star;
     private Date createDate;
 }
